Create maths tutorial(array)
In an online maths tutorial a student is given a list of N numbers. From this list, the student is required to sum each element in the list such that the ith element in the output list will be equal to the sum to the first element until the ith element in the list.

Input Format

The first line of input consists of an integer – numSize representing the count of total numbers in the list(N). The next line consists of N space separated integers. element1, element2,…… element-n representing the numeric value in the list.

Constraints